Move MavenArtifactIdentifier to m2e.jdt to make it more accessible
... without UI dependencies.
