+++ org.eclipse.m2e.jdt/META-INF/MANIFEST.MF
@@ -6,7 +6,8 @@ Bundle-Version: 2.0.2.qualifier
 Bundle-Localization: plugin
 Export-Package: org.eclipse.m2e.jdt,
  org.eclipse.m2e.jdt.internal;x-friends:="org.eclipse.m2e.jdt.ui",
- org.eclipse.m2e.jdt.internal.launch;x-friends:="org.eclipse.m2e.jdt.ui"
+ org.eclipse.m2e.jdt.internal.launch;x-friends:="org.eclipse.m2e.jdt.ui",
+ org.eclipse.m2e.sourcelookup.internal;x-friends:="org.eclipse.m2e.sourcelookup,org.eclipse.m2e.sourcelookup.ui"
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.debug.core,
  org.eclipse.jdt.core;bundle-version="3.18.0",
@@ -18,5 +19,7 @@ Bundle-ActivationPolicy: lazy
 Bundle-Activator: org.eclipse.m2e.jdt.MavenJdtPlugin
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-Vendor: %Bundle-Vendor
-Import-Package: org.slf4j;version="[1.7.0,3.0.0)"
+Import-Package: com.google.gson;version="[2.9.1,3.0.0)",
+ org.apache.commons.codec.digest;version="[1.14.0,2.0.0)",
+ org.slf4j;version="[1.7.0,3.0.0)"
 Automatic-Module-Name: org.eclipse.m2e.jdt
