+++ org.eclipse.m2e.jdt/src/org/eclipse/m2e/sourcelookup/internal/MetaInfMavenScanner.java
@@ -10,7 +10,7 @@
  * Contributors:
  *      Igor Fedorenko - initial API and implementation
  *******************************************************************************/
-package org.eclipse.m2e.sourcelookup.internal.launch;
+package org.eclipse.m2e.sourcelookup.internal;
 
 import java.io.File;
 import java.io.IOException;
