+++ org.eclipse.m2e.sourcelookup.ui/META-INF/MANIFEST.MF
@@ -9,12 +9,12 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.debug.ui;bundle-version="3.7.0",
  org.eclipse.m2e.core;bundle-version="[2.0.0,3.0.0)",
- org.eclipse.m2e.sourcelookup;bundle-version="[2.0.0,3.0.0)",
  org.eclipse.m2e.core.ui;bundle-version="[2.0.0,3.0.0)",
  org.eclipse.ui.ide;bundle-version="3.7.0",
  org.eclipse.m2e.binaryproject;bundle-version="[2.0.0,3.0.0)",
  org.eclipse.jdt.core;bundle-version="3.9.0",
- org.eclipse.jdt.launching;bundle-version="3.10.0"
+ org.eclipse.jdt.launching;bundle-version="3.10.0",
+ org.eclipse.m2e.jdt;bundle-version="[2.0.2,3.0.0)"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Automatic-Module-Name: org.eclipse.m2e.sourcelookup.ui
