+++ org.eclipse.m2e.sourcelookup/META-INF/MANIFEST.MF
@@ -10,8 +10,7 @@ Require-Bundle: org.eclipse.m2e.core;bundle-version="[2.0.0,3.0.0)",
  org.eclipse.jdt.launching;bundle-version="3.9.0",
  org.eclipse.core.runtime;bundle-version="3.7.0",
  org.eclipse.debug.ui;bundle-version="3.10.0",
- com.google.gson;bundle-version="2.2.4",
- org.apache.commons.codec;bundle-version="1.14.0"
+ org.eclipse.m2e.jdt;bundle-version="[2.0.2,3.0.0)"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Eclipse-BundleShape: dir
